Change `document_type` hardcodes to "public"
Previously, the only allowed `document_type`s for a case were `5766`
("Non-confidential"), and `5767` ("Confidential"). Tyler has changed
this language (which I'm now realizing always changes the code values
they represent) to be "Public" and "Non-public". I have changed the
search language to use the new terms, and changed the codes to the new
values. I kept some of the old codes when checking for things, because I
believe that the old codes are sometimes kept in older cases.